// frontend/src/api/generated/model/dashboard-status-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kafka Lab API
 * API for controlling and monitoring the Kafka lab demo.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ConsumerStatusResponse } from './consumer-status-response';
// May contain unused imports in some cases
// @ts-ignore
import type { LagStatusResponse } from './lag-status-response';
// May contain unused imports in some cases
// @ts-ignore
import type { ProducerStatusResponse } from './producer-status-response';
// May contain unused imports in some cases
// @ts-ignore
import type { RatesStatusResponse } from './rates-status-response';

export interface DashboardStatusResponse {
    /**
     * Time at which the snapshot was computed.
     */
    'timestamp': string;
    /**
     * Kafka topic monitored by the dashboard.
     */
    'topic': string;
    /**
     * Kafka consumer group monitored by the dashboard.
     */
    'consumerGroupId': string;
    'producer': ProducerStatusResponse;
    'consumer': ConsumerStatusResponse;
    'lag': LagStatusResponse;
    'rates': RatesStatusResponse;
}